% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{apply.quant.scale}
\alias{apply.quant.scale}
\title{Applying a quantile scale}
\usage{
apply.quant.scale(x, center, scale)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{center}{A centering vector for each column}

\item{scale}{A scaling vector for each column}
}
\value{
The centered, scaled matrix. The numeric centering and scalings used are returned as attributes.
}
\description{
Applying a quantile scale
}

