% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{apply_quantile_scale}
\alias{apply_quantile_scale}
\title{Apply a quantile scale.}
\usage{
apply_quantile_scale(x, addend, multiplier)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{addend}{A minimum vector for each column}

\item{multiplier}{A scaling vector for each column}
}
\value{
The scaled matrix. The numeric centering and scalings used are returned as attributes.
}
\description{
Anything outside the range of [0, 1] will be set to 0 or 1.
}
