% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{apply_scale}
\alias{apply_scale}
\title{Apply a uniform scale}
\usage{
apply_scale(x, center, scale)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{center}{A centering vector for each column}

\item{scale}{A scaling vector for each column}
}
\value{
The centered, scaled matrix. The numeric centering and scalings used are returned as attributes.
}
\description{
Apply a uniform scale
}
