% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluate_trajectory}
\alias{evaluate_trajectory}
\alias{evaluate_dim_red}
\title{Deprecated evaluation functions}
\usage{
evaluate_trajectory(time, progression)

evaluate_dim_red(space, progression, k = 5)
}
\arguments{
\item{time}{deprecated}

\item{progression}{deprecated}

\item{space}{deprecated}

\item{k}{deprecated}
}
\description{
These functions are outdated. Use \code{dyneval::evaluate_ti_method} instead. For more info, visit \url{https://dynverse.org}.
}
