% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{draw.trajectory.heatmap}
\alias{draw.trajectory.heatmap}
\title{Draw time-series heatmap}
\usage{
draw.trajectory.heatmap(
  x,
  time,
  progression.group=NULL,
  modules=NULL,
  show.labels.row=FALSE,
  show.labels.col=FALSE,
  scale.features=TRUE,
  narrow.breaks=TRUE,
  ...
)
}
\arguments{
\item{x}{A numeric matrix or data frame with one row per sample and one column per feature.}

\item{time}{A numeric vector containing the inferred time points of each sample along a trajectory.}

\item{progression.group}{\code{NULL} or a vector (or factor) containing the groupings of the samples (default \code{NULL}).}

\item{modules}{\code{NULL} or a data frame as returned by \code{\link{extract.modules}}.}

\item{show.labels.row}{\code{TRUE} if the labels of the rows are to be plotted (default \code{FALSE}).}

\item{show.labels.col}{\code{TRUE} if the labels of the cols are to be plotted (default \code{FALSE}).}

\item{scale.features}{\code{TRUE} if the values of each feature is to be scaled (default \code{TRUE}).}

\item{narrow.breaks}{\code{TRUE} if the breaks of the heatmap is to be adjust to allow a better visualisation (default \code{TRUE}).}

\item{...}{Optional arguments to \code{\link[pheatmap]{pheatmap}}}
}
\value{
The output of the \code{\link[pheatmap]{pheatmap}} function.
}
\description{
\code{draw.trajectory.heatmap} draws a heatmap in which the samples
are ranked according their position in an inferred trajectory. In addition, the progression groups and
feature modules can be passed along to further enhance the visualisation.
}
\examples{
## Generate a dataset
dataset <- generate.dataset(type="s", num.genes=500, num.samples=1000, num.groups=4)
expression <- dataset$expression
dist <- correlation.distance(expression)
space <- reduce.dimensionality(dist, ndim=2)
groups <- dataset$sample.info$group.name
traj <- infer.trajectory(space)
time <- traj$time

## Find the genes are most trajectory aligned
tafs <- find.trajectory.aligned.features(expression, time)
expr.tafs <- expression[,tafs$tafs]

## Draw a time series heatmap of the TAFs
draw.trajectory.heatmap(expr.tafs, time)

## Also show the progression groupings
draw.trajectory.heatmap(expr.tafs, time, progression=groups)

## Draw a heatmap of the smoothed data instead
smooth.tafs <- tafs$smooth.x[,tafs$tafs]
draw.trajectory.heatmap(smooth.tafs, time, progression=groups)

## Group the genes into modules and visualise the modules in a heatmap
modules <- extract.modules(smooth.tafs)
draw.trajectory.heatmap(expr.tafs, time, progression.group=groups, modules=modules)
}

