% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluate.dim.red}
\alias{evaluate.dim.red}
\title{Evaluate the dimensionality reduction}
\usage{
evaluate.dim.red(space, progression, k=5)
}
\arguments{
\item{space}{A numeric vector containing the inferred time points of each sample along a trajectory.}

\item{progression}{A factor or a numeric vector which represents the progression stages of each sample.}

\item{k}{The maximum number of nearest neighbours to search (default 5).}
}
\value{
The accuracy of a 5NN LOOCV using the dimensionality reduction to predict the progression stage of a sample.
}
\description{
\code{evaluate.dim.red} calculates the \emph{accuracy} of the
dimensionality reduction by performing 5-nearest neighbour leave-one-out-cross-validation (5NN LOOCV).
}
\examples{
## Generate a dataset
dataset <- generate.dataset(type="s", num.genes=500, num.samples=1000, num.groups=4)
dist <- correlation.distance(dataset$expression)
space <- reduce.dimensionality(dist, ndim=2)

## Evaluate the trajectory timeline
evaluate.dim.red(space, dataset$sample.info$group.name)
}

