% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_modules.R
\name{extract.modules}
\alias{extract.modules}
\title{Extract modules of features}
\usage{
extract.modules(x, ...)
}
\arguments{
\item{x}{A numeric matrix or data frame with \emph{M} rows (one per sample) and \emph{P} columns (one per feature).}

\item{...}{Extra parameters passed to Mclust}
}
\value{
A data frame containing meta-data for the features in \code{x}, namely the order in which to visualise the features in and which module they belong to.
}
\description{
\code{extract.modules} uses adaptive branch pruning to extract modules of features, which is typically done on the smoothed expression returned by \code{\link{gene.importances}}.
}
\examples{
## Generate a dataset and visualise
dataset <- generate.dataset(type="s", num.genes=500, num.samples=1000, num.groups=4)
expression <- dataset$expression
group.name <- dataset$sample.info$group.name
dist <- correlation.distance(expression)
space <- reduce.dimensionality(dist, ndim=2)
traj <- infer.trajectory(space)
time <- traj$time
draw.trajectory.plot(space, path=traj$path, group.name)

## Select most important genes
gimp <- gene.importances(expression, traj$time, num.permutations = 0)
gene.sel <- gimp[1:50,]
expr.sel <- expression[,gene.sel$gene]

## Group the genes into modules and visualise the modules in a heatmap
modules <- extract.modules(quant.scale(expr.sel))
draw.trajectory.heatmap(expr.sel, time, group.name, modules)
}
\seealso{
\code{\link{gene.importances}}
}
