% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_inference.R
\name{gene.importances}
\alias{gene.importances}
\title{Calculate the importance of a gene}
\usage{
gene.importances(x, time, ...)
}
\arguments{
\item{x}{A numeric matrix or data frame with \emph{M} rows (one per sample) and \emph{P} columns (one per feature).}

\item{time}{A numeric vector containing the inferred time points of each sample along a trajectory as returned by \code{\link{infer.trajectory}}.}

\item{...}{parameters passed to randomForest}
}
\description{
Calculate the importance of a gene
}

