% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_inference.R
\name{infer.consensus.trajectory}
\alias{infer.consensus.trajectory}
\title{Average orderings of multiple trajectories}
\usage{
infer.consensus.trajectory(space, number = 10, ...)
}
\arguments{
\item{space}{A numeric matrix or data frame containing the coordinates of samples.}

\item{number}{The number of trajectories to infer}

\item{...}{Parameters for \code{\link{infer.trajectory}}}
}
\value{
A list containing several objects:
\itemize{
  \item \code{path}: NULL
  \item \code{time}: the time point of each sample along the inferred trajectory.
}
}
\description{
\code{infer.consensus.trajectory} infers multiple
trajectories with the \code{\link{infer.trajectory}} method, and aligns
and averages the orderings.
}
\seealso{
\code{\link{infer.trajectory}}, \code{\link{reduce.dimensionality}}, \code{\link{draw.trajectory.plot}}
}
