% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_inference.R
\name{infer.guided.trajectory}
\alias{infer.guided.trajectory}
\title{Infer a guided trajectory}
\usage{
infer.guided.trajectory(space, number = 10, ...)
}
\arguments{
\item{space}{A numeric matrix or data frame containing the coordinates of samples.}

\item{number}{The number of trajectories to infer}

\item{...}{Parameters for \code{\link{infer.trajectory}}}
}
\value{
A list containing several objects:
\itemize{
  \item \code{path}: NULL
  \item \code{time}: the time point of each sample along the inferred trajectory.
}
}
\description{
\code{infer.guided.trajectory} infers multiple
trajectories and uses the original space with the combined trajectory orderings
to infer a guided trajectory.
}
\seealso{
\code{\link{infer.trajectory}}, \code{\link{reduce.dimensionality}}, \code{\link{draw.trajectory.plot}}
}

