% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_inference.R
\name{infer.initial.trajectory}
\alias{infer.initial.trajectory}
\title{Infer an initial trajectory through space}
\usage{
infer.initial.trajectory(space, k)
}
\arguments{
\item{space}{A numeric matrix or data frame containing the coordinates of samples.}

\item{k}{The number of clusters to cluster the data into.}
}
\value{
the initial trajectory obtained by this method
}
\description{
\code{infer.initial.trajectory} infers an initial trajectory for
\code{\link{infer.trajectory}} by clustering the points and calculating
the shortest path through cluster centers. The shortest path takes into account
the euclidean distance between cluster centers, and the density between those two
points.
}
\examples{
## Generate an example dataset and visualise it
dataset <- generate.dataset(type = "poly", num.genes = 500, num.samples = 1000, num.groups = 4)
dist <- correlation.distance(dataset$expression)
space <- reduce.dimensionality(dist, ndim = 2)
draw.trajectory.plot(space, progression.group = dataset$sample.info$group.name)

## Infer a trajectory through this space
init.traj <- infer.initial.trajectory(space)

## Visualise the trajectory
draw.trajectory.plot(space, path = init.traj, progression.group = dataset$sample.info$group.name)
}

