% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{infer.random.trajectory}
\alias{infer.random.trajectory}
\title{Infer random trajectory through space}
\usage{
infer.random.trajectory(space, k)
}
\arguments{
\item{space}{A numeric matrix or data frame.}

\item{k}{The number of clusters to cluster the data into.}
}
\value{
A list of the same structure as \link{\code{infer.trajectory}}
}
\description{
\code{infer.random.trajectory} infers a random trajectory in the given space.
}
\examples{
## generate example dataset and visualise it
x <- seq(-1, 1, by=.002)
group <- cut(x, breaks=4)
y <- rowSums(poly(x, 3))*4
space <- cbind(Comp1=x + rnorm(length(x), sd=.1), Comp2=y + rnorm(length(y), sd=.2))
space <- rescale.and.center(space)
plot.dimensionality.reduction(space, colour=group)

## infer trajectory and visualise
traj <- infer.random.trajectory(space, k=10)
plot.trajectory(space, traj$final.path, colour=group)
}

