% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_inference.R
\name{infer.trajectory}
\alias{infer.trajectory}
\title{Infer linear trajectory through space}
\usage{
infer.trajectory(space, k = 4, thresh = .001, maxit = 10, stretch = 0, smoother = "smooth.spline")
}
\arguments{
\item{space}{A numeric matrix or data frame containing the coordinates of samples.}

\item{k}{The number of clusters to cluster the data into.}

\item{thresh}{\code{\link[princurve]{principal.curve}} parameter: convergence threshhold on shortest distances to the curve}

\item{maxit}{\code{\link[princurve]{principal.curve}} parameter: maximum number of iterations}

\item{stretch}{\code{\link[princurve]{principal.curve}} parameter: a factor by which the curve can be extrapolated when points are projected}

\item{smoother}{\code{\link[princurve]{principal.curve}} parameter: choice of smoother}
}
\value{
A list containing several objects:
\itemize{
  \item \code{path}: the trajectory obtained by principal curves.
  \item \code{time}: the time point of each sample along the inferred trajectory.
}
}
\description{
\code{infer.trajectory} infers a trajectory through samples in a given space in a four-step process:
\enumerate{
  \item Perform \emph{k}-means clustering
  \item Calculate distance matrix between cluster centers using a custom distance function
  \item Find the shortest path connecting all cluster centers using the custom distance matrix
  \item Iteratively fit a curve to the given data using principal curves
}
}
\examples{
## Generate an example dataset and visualise it
dataset <- generate.dataset(type="poly", num.genes=500, num.samples=1000, num.groups=4)
dist <- correlation.distance(dataset$expression)
space <- reduce.dimensionality(dist, ndim=2)
draw.trajectory.plot(space, progression.group=dataset$sample.info$group.name)

## Infer a trajectory through this space
traj <- infer.trajectory(space)

## Visualise the trajectory
draw.trajectory.plot(space, path=traj$path, progression.group=dataset$sample.info$group.name)
}
\seealso{
\code{\link{reduce.dimensionality}}, \code{\link{draw.trajectory.plot}}
}
