% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_inference.R
\name{infer_initial_trajectory}
\alias{infer_initial_trajectory}
\title{Infer an initial trajectory through space}
\usage{
infer_initial_trajectory(space, k)
}
\arguments{
\item{space}{A numeric matrix or a data frame containing the coordinates of samples.}

\item{k}{The number of clusters}
}
\value{
the initial trajectory obtained by this method
}
\description{
\code{infer_initial_trajectory} infers an initial trajectory for
\code{\link{infer_trajectory}} by clustering the points and calculating
the shortest path through cluster centers. The shortest path takes into account
the euclidean distance between cluster centers, and the density between those two
points.
}
