% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\name{knn.distances}
\alias{knn.distances}
\title{k Nearest Neighbour distances}
\usage{
knn.distances(dist, k, self.loops=F)
}
\arguments{
\item{dist}{A numeric matrix, data frame or "\code{dist}" object.}

\item{k}{The maximum number of nearest neighbours to search.}

\item{self.loops}{\code{TRUE} if samples with the same index or name are allowed to be neighbours.}
}
\value{
A matrix containing the distances of the \emph{k} nearest neighbours of each sample.
}
\description{
\code{knn.distances} returns the distances of the \emph{k} nearest neighbours of each sample.
}
\examples{
## Calculate the kNN distances within a set of samples
x <- matrix(rnorm(50*10, mean=0, sd=1), ncol=10)
dist <- dist(x)
knnd <- knn.distances(dist, 10)
plot(density(knnd))

## Calculate the kNN distances between two sets of samples
y <- matrix(rnorm(100*10, mean=1, sd=2), ncol=10)
dist <- euclidean.distance(x, y)
knnd <- knn.distances(dist, 10)
plot(density(knnd))
}
