% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{landmark_selection}
\alias{landmark_selection}
\title{Select landmarks}
\usage{
landmark_selection(x, dist_fun, landmark_method, num_landmarks)
}
\arguments{
\item{x}{a numeric matrix}

\item{dist_fun}{the distance function to be used; must have exactly two arguments, namely dist_fun(x, y).}

\item{landmark_method}{Must be "naive" for now. Other landmark methods will be supported in the future.}

\item{num_landmarks}{the number of landmarks to be selected.}
}
\value{
A list with the following elements:
\itemize{
  \item{\code{ix_lm}: The incides of the selected landmarks}
  \item{\code{dist_lm}: Pairwise distance matrix between the selected landmarks}
  \item{\code{dist_2lm}: Distance matrix between the landmarks and all the samples in \code{x}}
}
}
\description{
Select landmarks
}
