% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{outlier.filter}
\alias{outlier.filter}
\title{Outlier detection}
\usage{
outlier.filter(dist)
}
\arguments{
\item{dist}{A numeric matrix, data frame or "\code{dist}" object.}
}
\value{
A boolean vector indicating whether samples are \emph{not} outliers.
}
\description{
\code{outlier.filter} calculates which samples are outliers by iteratively
removing the samples with the highest \emph{outlierness}' and fitting a normal distribution
to the remaining outlierness values. A selection of samples is made by picking the iteration
at which the log likelihood is maximised.
}
\examples{
## Generate normally distributed points, calculate their outliernesses and which points are outliers
x <- matrix(rnorm(200*2), ncol=2)
dist <- euclidean.distance(x)
filt <- outlier.filter(dist)
# plot points using their outlierness value as size and whether or not they were outliers as colours
plot(x, col=filt+2, cex=outlierness(dist)+1, pch=20)
# plot the score at each iteration of the removal process
likelihood.df <- attr(filt, "loglikelihood")
plot(likelihood.df$amount.removed, likelihood.df$log.likelihood, type="l")

## Generate a random expression dataset
dataset <- generate.dataset(type="poly", num.genes=500, num.samples=200, num.groups=4)
dist <- correlation.distance(dataset$expression)
space <- reduce.dimensionality(dist, ndim=2)
filt <- outlier.filter(dist)
# plot points using their outlierness value as size and whether or not they were outliers as colours
plot(space, col=filt+2, cex=outlierness(dist)+1, pch=20)
# plot the score at each iteration of the removal process
likelihood.df <- attr(filt, "loglikelihood")
plot(likelihood.df$amount.removed, likelihood.df$log.likelihood, type="l")
}

