% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{plot.dimensionality.reduction}
\alias{plot.dimensionality.reduction}
\title{Title}
\usage{
plot.dimensionality.reduction(space, colour = NULL, contour = F)
}
\arguments{
\item{space}{}

\item{colour}{}

\item{contour}{}
}
\description{
Title
}

