% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{plot.modules.heatmap}
\alias{plot.modules.heatmap}
\title{Title}
\usage{
plot.modules.heatmap(smooth.expression, tag.genes, progression.str, time,
  modules)
}
\arguments{
\item{progression.str}{}

\item{time}{}

\item{modules}{}

\item{tags}{}
}
\description{
Title
}

