% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.scorpius}
\alias{plot.scorpius}
\title{Visualise SCORPIUS}
\usage{
plot.scorpius(space, progression.group=NULL, path=NULL, contour=F)
}
\arguments{
\item{space}{A numeric matrix or data frame containing the coordinates of samples.}

\item{progression.group}{\code{NULL} or a vector (or factor) containing the groupings of the samples (default \code{NULL}).}

\item{path}{A numeric matrix or data frame containing the coordinates of the inferred path.}

\item{contour}{\code{TRUE} if contours are to be drawn around the samples. This will only work with ggplot2 (≥ 2.0).}
}
\value{
A ggplot2 plot.
}
\description{
\code{plot.scorpius} is used to plot samples after performing dimensionality reduction. 
Additional arguments can be provided to colour the samples, plot the trajectory inferred by SCORPIUS,
and draw a contour around the samples.
}
\examples{
## Generate a synthetic dataset
dataset <- generate.dataset(type="p", num.genes=500, num.samples=1000, num.groups=4)
dist <- correlation.distance(dataset$expression)
space <- reduce.dimensionality(dist, ndim=2)
groups <- dataset$sample.info$group.name

## Simply plot the samples
plot.scorpius(space)

## Colour each sample according to its group
plot.scorpius(space, progression.group=groups)

## Add contours to the plot
plot.scorpius(space, progression.group=groups, contour=T)

## Plot contours without colours
plot.scorpius(space, progression.group=T)

## Infer a trajectory and plot it
traj <- infer.trajectory(space, k=4)
plot.scorpius(space, progression.group=groups, path=traj$final.path)
plot.scorpius(space, progression.group=groups, path=traj$final.path, contour=T)
}

