% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{plot.trajectory.density}
\alias{plot.trajectory.density}
\title{Title}
\usage{
plot.trajectory.density(time, progression, colour = NULL)
}
\arguments{
\item{time}{}

\item{progression}{}

\item{colour}{}
}
\description{
Title
}

