% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.trajectory.heatmap}
\alias{plot.trajectory.heatmap}
\title{Title}
\usage{
plot.trajectory.heatmap(x, time, progression = NULL, modules = NULL,
  show.labels.row = F, show.labels.col = F, scale.features = T,
  narrow.breaks = T)
}
\arguments{
\item{x}{}

\item{time}{}

\item{progression}{}

\item{modules}{}

\item{show.labels.row}{}

\item{show.labels.col}{}

\item{scale.features}{}
}
\description{
Title
}

