% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{plot.trajectory}
\alias{plot.trajectory}
\title{Title}
\usage{
plot.trajectory(space, path, colour = NULL, contour = F)
}
\arguments{
\item{space}{}

\item{path}{}

\item{colour}{}
}
\description{
Title
}

