% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction.R
\name{quant.scale}
\alias{quant.scale}
\title{Calculating and applying a quantile scale}
\usage{
quant.scale(x, outlier.cutoff = 0.05)
}
\arguments{
\item{x}{A numeric matrix or data frame.}

\item{outlier.cutoff}{The quantile cutoff for outliers (default 0.05).}
}
\value{
The centered, scaled matrix. The numeric centering and scalings used are returned as attributes.
}
\description{
Calculating and applying a quantile scale
}
\examples{
## Generate a matrix from a normal distribution
## with a large standard deviation, centered at c(5, 5)
x <- matrix(rnorm(200*2, sd = 10, mean = 5), ncol=2)

## Center the dataset at c(0, 0) with a minimum of c(-.5, -.5) and a maximum of c(.5, .5)
x.scaled <- quant.scale(x)

## Plot rescaled data
plot(x.scaled)

## Show ranges of each column
apply(x.scaled, 2, range)
}

