% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction_landmarked.R
\name{reduce.dimensionality.landmarked}
\alias{reduce.dimensionality.landmarked}
\title{Multidimensional scaling with landmarks}
\usage{
reduce.dimensionality.landmarked(x, dist.fun, ndim = 3,
  landmark.method = "naive", num.landmarks = 1000, rescale = T)
}
\arguments{
\item{x}{a numeric matrix}

\item{dist.fun}{the distance function to be used; must have exactly two arguments, namely dist.fun(x, y).}

\item{ndim}{the maximum dimension of the space which the data are to be represented in; must be in {1, 2, \ldots, n-1}.}

\item{landmark.method}{Must be "naive" for now. Other landmark methods will be supported in the future.}

\item{num.landmarks}{the number of landmarks to be selected.}

\item{rescale}{A logical indicating whether or not the returned space should be rescaled and centered.}
}
\value{
A list containing the reduced space of the landmarks and the complete dataset.
}
\description{
Multidimensional scaling with landmarks
}
\examples{
data(ginhoux)
space <- reduce.dimensionality.landmarked(
  ginhoux$expression,
  correlation.distance,
  num.landmarks = 200)
draw.trajectory.plot(space, ginhoux$sample.info$group.name)
}
