% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{reduce.dimensionality}
\alias{reduce.dimensionality}
\title{Title}
\usage{
reduce.dimensionality(dist, ndim, scaled = T)
}
\arguments{
\item{ndim}{}

\item{scaled}{}

\item{dis}{}
}
\description{
Title
}

