% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{rescale.and.center}
\alias{rescale.and.center}
\title{Title}
\usage{
rescale.and.center(space, center = 0, max.range = 1)
}
\arguments{
\item{space}{}

\item{center}{}

\item{max.range}{}
}
\description{
Title
}

