% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{SCORPIUS-deprecated}
\alias{SCORPIUS-deprecated}
\alias{generate.dataset}
\alias{reduce.dimensionality}
\alias{reduce.dimensionality.landmarked}
\alias{euclidean.distance}
\alias{correlation.distance}
\alias{knn.distances}
\alias{evaluate.dim.red}
\alias{evaluate.trajectory}
\alias{extract.modules}
\alias{gene.importances}
\alias{outlier.filter}
\alias{draw.trajectory.plot}
\alias{draw.trajectory.heatmap}
\alias{rescale.and.center}
\alias{apply.scale}
\alias{quant.scale}
\alias{apply.quant.scale}
\alias{infer.initial.trajectory}
\alias{infer.trajectory}
\alias{reverse.trajectory}
\alias{SCORPIUS-deprecated-package}
\title{Deprecated function(s) in the SCORPIUS package}
\usage{
generate.dataset(...)
}
\arguments{
\item{...}{Parameters to be passed to the new version of the function}
}
\description{
These functions are provided for compatibility with older version of
the SCORPIUS package.  They may eventually be completely removed.
}
\section{Details}{

\tabular{rl}{
  \code{generate.dataset} \tab now a synonym for \code{\link{generate_dataset}}\cr
}
}

