% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynwrap_integration.R
\name{ti_scorpius_run_fun}
\alias{ti_scorpius_run_fun}
\title{Run scorpius using the dynwrap pipeline}
\usage{
ti_scorpius_run_fun(expression, priors, parameters, seed = NULL, verbose = 0)
}
\arguments{
\item{expression}{Expression matrix}

\item{priors}{Priors}

\item{parameters}{Parameters}

\item{seed}{Random seed}

\item{verbose}{Verbosity level}
}
\description{
Run scorpius using the dynwrap pipeline
}
