% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_inference.R
\name{find.trajectory.aligned.features}
\alias{find.trajectory.aligned.features}
\title{Find trajectory-aligned features}
\usage{
find.trajectory.aligned.features(x, time, p.adjust.method="BH", q.value.cutoff=1e-10, df=8, parallel=F)
}
\arguments{
\item{x}{A numeric matrix or data frame with \emph{M} rows (one per sample) and \emph{P} columns (one per feature).}

\item{time}{A numeric vector containing the inferred time points of each sample along a trajectory as returned by \code{\link{infer.trajectory}}.}

\item{p.adjust.method}{The correction method used by \code{\link[stats]{p.adjust}}.}

\item{q.value.cutoff}{The cutoff used on the q-values.}

\item{df}{The degree of freedom used by \code{\link[splines]{ns}}.}

\item{parallel}{Must be \code{FALSE} (default), \code{TRUE} (will auto-detect number of cores), or a numeric value specifying the number of cores to use.}
}
\value{
Returns a list, containing the following items: \itemize{
  \item \code{tafs}: the names or indices of all trajectory-aligned features (q-value < \code{q.value.cutoff}),
  \item \code{p.values}: a data frame containing an adjusted p-value for each feature,
  \item \code{smooth.x}: the input matrix \code{x} smoothed by splines.
}
}
\description{
\code{find.trajectory.aligned.features} searches for features whose values . It evaluates how well splines are able model the values of a given feature over the timeline of a trajectory.
}
\examples{
## Generate a dataset and visualise
dataset <- generate.dataset(type="s", num.genes=500, num.samples=1000, num.groups=4)
expression <- dataset$expression
dist <- correlation.distance(expression)
space <- reduce.dimensionality(dist, ndim=2)
traj <- infer.trajectory(space, k=4)
draw.trajectory.plot(space, path=traj$final.path, progression.group=dataset$sample.info$group.name)

## Show which genes are most trajectory aligned
tafs <- find.trajectory.aligned.features(expression, traj$time)
head(tafs$p.values, 10)

## Visualise the expression of these genes in a heatmap
expr.tafs <- expression[,tafs$tafs]
draw.trajectory.heatmap(expr.tafs, time=traj$time, progression.group=dataset$sample.info$group.name)
}
\seealso{
\code{\link{infer.trajectory}}, \code{\link{draw.trajectory.heatmap}}
}

