% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{find.trajectory.aligned.genes}
\alias{find.trajectory.aligned.genes}
\title{Title}
\usage{
find.trajectory.aligned.genes(counts, time, degrees.of.freedom = 8,
  p.adjust.method = "BH", q.value.cutoff = 1e-10, mc.cores = 1)
}
\arguments{
\item{counts}{}

\item{time}{}

\item{degrees.of.freedom}{}

\item{p.adjust.method}{}
}
\description{
Title
}

