% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_importances.R
\name{gene.importances}
\alias{gene.importances}
\title{Calculate the importance of a feature}
\usage{
gene.importances(x, time, num.permutations = 0, ntree = 10000,
  mtry = ncol(x) * 0.01, num.threads = NULL, ...)
}
\arguments{
\item{x}{A numeric matrix or data frame with \emph{M} rows (one per sample) and \emph{P} columns (one per feature).}

\item{time}{A numeric vector containing the inferred time points of each sample along a trajectory as returned by \code{\link{infer.trajectory}}.}

\item{num.permutations}{The number of permutations to test against for calculating the p-values (default: 0).}

\item{ntree}{The number of trees to grow (default: 10000).}

\item{mtry}{The number of variables randomly samples at each split (default: 1\% of features).}

\item{num.threads}{Number of threads. Default is number of CPU cores available.}

\item{...}{Extra parameters passed to ranger.}
}
\value{
a data frame containing the importance of each feature for the given time line
}
\description{
Calculates the feature importance of each column in \code{x} in trying to predict the time ordering.
}
\examples{
dataset <- generate.dataset(type="s", num.genes=500, num.samples=1000, num.groups=4)
expression <- dataset$expression
group.name <- dataset$sample.info$group.name
dist <- correlation.distance(expression)
space <- reduce.dimensionality(dist, ndim=2)
traj <- infer.trajectory(space)
gene.importances(expression, traj$time, num.permutations = 0)
}
