% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{generate.dataset}
\alias{generate.dataset}
\title{Generate a synthetic dataset}
\usage{
generate.dataset(type=c("splines", "polynomial"), num.samples=1000, num.genes=100, num.groups=4)
}
\arguments{
\item{type}{The type of function used in order to generate the expression data. Must be either \code{"splines"} (default) or \code{"polynomial"} (or abbreviations thereof).}

\item{num.samples}{The number of samples the dataset will contain.}

\item{num.genes}{The number of genes the dataset will contain.}

\item{num.groups}{The number of groups the samples will be split up in.}
}
\value{
A list containing the expression data and the meta data of the samples.
}
\description{
\code{generate.dataset} generates an synthetic dataset which can be used for visualisation purposes.
}
\examples{
## Generate a dataset
dataset <- generate.dataset(type="poly", num.genes=500, num.samples=1000, num.groups=4)

## Reduce dimensionality and infer trajectory with SCORPIUS
dist <- correlation.distance(dataset$expression)
space <- reduce.dimensionality(dist, ndim=2)
traj <- infer.trajectory(space, k=4)

## Visualise
plot.scorpius(space, path=traj$final.path, colour=dataset$sample.info$group.name)
}

