% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{infer.trajectory}
\alias{infer.trajectory}
\title{Infer linear trajectory through space}
\usage{
infer.trajectory(space, k)
}
\arguments{
\item{space}{A numeric matrix or data frame.}

\item{k}{The number of clusters to cluster the data into.}
}
\value{
A list containing several objects:
\itemize{
  \item \code{clustering}: the initial \emph{k}-means clustering.
  \item \code{initial.path}: the initial shortest path through the different cluster centers.
  \item \code{final.path}: the final trajectory obtained by principal curves.
  \item \code{time}: the time point of each sample along the inferred trajectory.
}
}
\description{
\code{infer.trajectory} infers a trajectory through samples in a given space in a four-step process:
\enumerate{
  \item Perform \emph{k}-means clustering
  \item Calculate distance matrix between cluster centers using a custom distance function
  \item Find the shortest path connecting all cluster centers using the custom distance matrix
  \item Iteratively fit a curve to the given data using principal curves 
}
}
\examples{
## Generate an example dataset and visualise it
dataset <- generate.dataset(type="poly", num.genes=500, num.samples=1000, num.groups=4)
dist <- correlation.distance(dataset$expression)
space <- reduce.dimensionality(dist, ndim=2)
plot.scorpius(space, progression.group=dataset$sample.info$group.name)

## Infer a trajectory through this space
traj <- infer.trajectory(space, k=4)

## Visualise the trajectory
plot.scorpius(space, path=traj$final.path, progression.group=dataset$sample.info$group.name)
}

