% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{outlier.filter}
\alias{outlier.filter}
\title{Outlier detection}
\usage{
outlier.filter(dist)
}
\arguments{
\item{dist}{A numeric matrix, data frame or "\code{dist}" object.}
}
\value{
A boolean vector indicating whether samples are \emph{not} outliers.
}
\description{
\code{outlier.filter} calculates which samples are outliers by iteratively
removing the samples with the highest \emph{outlierness}' and fitting a normal distribution
to the remaining outlierness values. A selection of samples is made by picking the iteration
at which the log likelihood is maximised.
}
\examples{
## generate uniformily distributed points, calculate their outliernesses and which points are outliers
x <- matrix(runif(200*2), ncol=2)
dist <- euclidean.distance(x)
filt <- outlier.filter(dist)
plot(x, col=filt+2, cex=outlierness(dist)+1, pch=20)

## generate normally distributed points, calculate their outliernesses and which points are outliers
x <- matrix(rnorm(200*2), ncol=2)
dist <- euclidean.distance(x)
filt <- outlier.filter(dist)
plot(x, col=filt+2, cex=outlierness(dist)+.5, pch=20)
}

