% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{reduce.dimensionality}
\alias{reduce.dimensionality}
\title{Dimensionality reduction}
\usage{
reduce.dimensionality(dist, ndim, rescale=TRUE)
}
\arguments{
\item{dist}{A numeric matrix, data frame or "\code{dist}" object.}

\item{ndim}{The number of dimensions in the new space.}

\item{rescale}{A logical indicating whether or not the returned space should be rescaled and centered.}
}
\value{
A matrix containing the coordinates of each sample, represented in an \code{ndim}-dimensional space.
}
\description{
\code{reduce.dimensionality} performs an eigenanalysis of the given dissimilarity matrix and returns coordinates of the samples represented in an \code{ndim}-dimensional space.
}
\examples{
## Generate an example dataset
dataset <- generate.dataset(type="poly", num.genes=500, num.samples=1000, num.groups=4)

## Reduce the dimensionality of this dataset
dist <- correlation.distance(dataset$expression)
space <- reduce.dimensionality(dist, ndim=2)

## Visualise the dataset
draw.trajectory.plot(space, progression.group=dataset$sample.info$group.name)
}
\seealso{
\code{\link{correlation.distance}}, \code{\link{euclidean.distance}}, \code{\link{rescale.and.center}}, \code{\link{draw.trajectory.plot}}
}

