% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{scale_quantile}
\alias{scale_minmax}
\alias{scale_uniform}
\alias{apply_quantile_scale}
\alias{apply_minmax_scale}
\alias{apply_uniform_scale}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dynutils}{\code{\link[dynutils]{scale_quantile}}, \code{\link[dynutils]{scale_minmax}}, \code{\link[dynutils]{scale_uniform}}, \code{\link[dynutils]{apply_quantile_scale}}, \code{\link[dynutils]{apply_minmax_scale}}, \code{\link[dynutils]{apply_uniform_scale}}}
}}

