% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{euclidean_distance}
\alias{reexports}
\alias{correlation_distance}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dynutils}{\code{\link[dynutils]{euclidean_distance}}, \code{\link[dynutils]{correlation_distance}}}
}}

