% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorpius.R
\name{reverse.trajectory}
\alias{reverse.trajectory}
\title{Reverse a trajectory}
\usage{
reverse.trajectory(trajectory)
}
\arguments{
\item{trajectory}{A trajectory as returned by \code{\link{infer.trajectory}}.}
}
\value{
The same trajectory, but in the other direction.
}
\description{
Since the direction of the trajectory is not specified, the ordering of a trajectory may be inverted using \code{reverse.trajectory}.
}
\examples{
## Generate an example dataset and infer a trajectory through it
dataset <- generate.dataset(type="poly", num.genes=500, num.samples=1000, num.groups=4)
dist <- correlation.distance(dataset$expression)
space <- reduce.dimensionality(dist, ndim=2)
traj <- infer.trajectory(space, k=4)

## Visualise the trajectory
draw.trajectory.plot(space, path=traj$final.path, progression.group=dataset$sample.info$group.name)

## Reverse the trajectory
reverse.traj <- reverse.trajectory(traj)
draw.trajectory.plot(space, path=reverse.traj$final.path, progression.group=dataset$sample.info$group.name)

## It's the same but reversed?!
plot(traj$time, reverse.traj$time, type="l")
}
\seealso{
\code{\link{infer.trajectory}}
}

